[BUG-17] Fixing app crashing when opening second file
Binding the DBPathAndPassword struct using fyne binding.UnTyped throws an error on the 2nd time the binding variable is set because the struct is not comparable due to the []byte field. To get around that I used a bound string to notify the DataView of changes and shared a single instance of DBPathAndPassword with a pointer.